pyproject.toml
[build-system]
requires = ["flit_core >=2,<4"]
build-backend = "flit_core.buildapi"

[project]
name =  "REvoDesign"
authors = [
    {name = "Yinying Yao", email = "yaoyy.hi@gmail.com"},
]
description = "A PyMOL plugin that makes enzyme redesign tasks easier to all."
readme = "README.md"
classifiers = [
    "Topic :: Scientific/Engineering :: Biochemistry",
    "Topic :: Scientific/Engineering :: Protein Engineering",
    "Development Status :: 6 - Mature",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dynamic = ["version"]

dependencies = [
    "typing-extensions>=4.9.0",
    "immutabledict>=4.1.0",
    "biopython>=1.81",
    "joblib>=1.3.2",
    "matplotlib>=3.6.0",
    "pandas>=2.1.0",
    "requests>=2.31.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.2",
    "pyOpenSSL>=23.3.0",
    "pybiolib>=1.1.1666",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "rdkit>=2023.9.4",
    "msgpack",
    "RosettaPy>=0.2.6",
    "bibtexparser @ git+https://github.com/sciunto-org/python-bibtexparser.git@v2.0.0b7#egg=bibtexparser",
    "platformdirs",
]

[project.optional-dependencies]
colabdesign = [
    "colabdesign @ git+https://github.com/sokrypton/ColabDesign.git@v1.1.1#egg=colabdesign",
]

pippack = [
    "PIPPack @ git+https://github.com/YaoYinYing/PIPPack.git@pip-installable-cpu#egg=PIPPack",
]

dlpacker = [
    "DLPacker @ git+https://github.com/YaoYinYing/DLPacker.git@pip-installable-cpu#egg=DLPacker",
]

spark = [
    "pyspark>=3.0.0"
]


test = [
    
    "bandit[toml]==1.7.10",
    "black<=24.10.0",
    "check-manifest<=0.50",
    "flake8-bugbear==24.8.19",
    "flake8-docstrings",
    "flake8-formatter_junit_xml",
    "flake8",
    "flake8-pyproject",
    "pylint==3.2.7",
    "pylint_junit",
    "pytest",
    "pytest-qt>=4.4.0",
    "pytest-cov<=6.0.0",
    "pytest-mock<3.14.1",
    "pytest-runner",
    "pytest<=8.3.3",
    "pytest-order",
    "pytest-github-actions-annotate-failures",
    "shellcheck-py==0.10.0.1",
    "pre-commit",
    "virtualenv<=20.27.1",
    "pooch>=1.8.0",
    "coverage",
]


[project.urls]
Documentation = "https://github.com/YaoYinYing/REvoDesign/tree/main#readme"
Source = "https://github.com/YaoYinYing/REvoDesign"
Tracker = "https://github.com/YaoYinYing/REvoDesign/issues"


[tool.flit.module]
name = "REvoDesign"

[tool.bandit]
exclude_dirs = ["build","dist","tests","scripts"]
number = 4
recursive = true
targets = "src"

[tool.black]
line-length = 120
fast = true

[tool.coverage.run]
branch = false

[tool.coverage.report]
fail_under = 100

[tool.flake8]
max-line-length = 120
select = "F,E,W,B,B901,B902,B903"
exclude = [
    ".eggs",
    ".git",
    ".tox",
    "nssm",
    "obj",
    "out",
    "packages",
    "pywin32",
    "tests",
    "swagger_client"
]
ignore = [
    "E722",
    "B001",
    "W503",
    "E203"
]

[tool.pyright]
include = ["src"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
]
venv = "env39"

reportMissingImports = true
reportMissingTypeStubs = false

pythonVersion = "3.9"
pythonPlatform = "Linux"

executionEnvironments = [
  { root = "src" }
]

[tool.pytest.ini_options]
addopts = "--cov-report xml:coverage.xml --cov src --cov-fail-under 0 --cov-append -m 'not integration' -v --pyargs --durations=0"
pythonpath = [
  "src"
]
testpaths = "tests"
junit_family = "xunit2"
markers = [
    "integration: marks as integration test",
    "notebooks: marks as notebook test",
    "gpu: marks as gpu test",
    "spark: marks tests which need Spark",
    "slow: marks tests as slow",
    "unit: fast offline tests",
]
filterwarnings = [
    "ignore::Bio.PDB.PDBExceptions.PDBConstructionWarning",
    ]


[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py, integration, spark, all

[testenv]
commands =
    pytest -m "not integration and not spark" {posargs}

[testenv:integration]
commands =
    pytest -m "integration" {posargs}

[testenv:spark]
extras = spark
setenv =
       PYSPARK_DRIVER_PYTHON = {envpython}
       PYSPARK_PYTHON = {envpython}
commands =
    pytest -m "spark" {posargs}

[testenv:all]
extras = all
setenv =
       PYSPARK_DRIVER_PYTHON = {envpython}
       PYSPARK_PYTHON = {envpython}
commands =
    pytest {posargs}
"""

[tool.pylint]
extension-pkg-whitelist= [
    "numpy",
    "torch",
    "cv2",
    "pyodbc",
    "pydantic",
    "ciso8601",
    "netcdf4",
    "scipy"
]
ignore="CVS"
ignore-patterns="test.*?py,conftest.py,test_*?py,TestData.py,*?Test.py"
limit-inference-results=100

[tool.pylint.'MESSAGES CONTROL']
enable="c-extension-no-member"

[tool.pylint.'REPORTS']

[tool.pylint.'REFACTORING']
max-nested-blocks=5
never-returning-functions="sys.exit"

[tool.pylint.'BASIC']
argument-naming-style="snake_case"
attr-naming-style="snake_case"
bad-names= [
    "foo",
    "bar"
]
class-attribute-naming-style="any"
class-naming-style="PascalCase"
const-naming-style="UPPER_CASE"
docstring-min-length=-1
function-naming-style="snake_case"
good-names= [
    "i",
    "j",
    "k",
    "ex",
    "Run",
    "_"
]
include-naming-hint="yes"
inlinevar-naming-style="any"
method-naming-style="snake_case"
module-naming-style="any"
no-docstring-rgx="^_"
property-classes="abc.abstractproperty"
variable-naming-style="snake_case"

[tool.pylint.'FORMAT']
ignore-long-lines="^\\s*(# )?.*['\"]?<?https?://\\S+>?"
indent-after-paren=4
indent-string='    '
max-line-length=120
max-module-lines=1000
single-line-class-stmt="no"
single-line-if-stmt="no"

[tool.pylint.'LOGGING']
logging-format-style="old"
logging-modules="logging"

[tool.pylint.'MISCELLANEOUS']
notes= [
    "FIXME",
    "XXX",
    "TODO"
]

[tool.pylint.'SIMILARITIES']
ignored-classes="Ui_Dialog,QtProgressBarHint"
ignore-comments="yes"
ignore-docstrings="yes"
ignore-imports="yes"
min-similarity-lines=7

[tool.pylint.'SPELLING']
max-spelling-suggestions=4
spelling-store-unknown-words="no"

[tool.pylint.'STRING']
check-str-concat-over-line-jumps="no"

[tool.pylint.'TYPECHECK']
contextmanager-decorators="contextlib.contextmanager"
generated-members="numpy.*,np.*,pyspark.sql.functions,collect_list"
ignore-none="yes"
ignore-on-opaque-inference="yes"
ignored-classes="optparse.Values,thread._local,_thread._local,numpy,torch,swagger_client,Ui_Dialog,QtProgressBarHint,"
ignored-modules="numpy,torch,swagger_client,netCDF4,scipy,Ui_REvoDesign-PyMOL-entry"
missing-member-hint="yes"
missing-member-hint-distance=1
missing-member-max-choices=1

[tool.pylint.'VARIABLES']
additional-builtins="dbutils"
allow-global-unused-variables="yes"
callbacks= [
    "cb_",
    "_cb"
]
dummy-variables-rgx="_+$|(_[a-zA-Z0-9_]*[a-zA-Z0-9]+?$)|dummy|^ignored_|^unused_"
ignored-argument-names="_.*|^ignored_|^unused_"
init-import="no"
redefining-builtins-modules="six.moves,past.builtins,future.builtins,builtins,io"

[tool.pylint.'CLASSES']
ignored-classes="Ui_Dialog,QtProgressBarHint"
defining-attr-methods= [
    "__init__",
    "__new__",
    "setUp",
    "__post_init__"
]
exclude-protected= [
    "_asdict",
    "_fields",
    "_replace",
    "_source",
    "_make"
]
valid-classmethod-first-arg="cls"
valid-metaclass-classmethod-first-arg="cls"

[tool.pylint.'DESIGN']
max-args=5
max-attributes=7
max-bool-expr=5
max-branches=12
max-locals=15
max-parents=7
max-public-methods=20
max-returns=6
max-statements=50
min-public-methods=2

[tool.pylint.'IMPORTS']
allow-wildcard-with-all="no"
analyse-fallback-blocks="no"
deprecated-modules="optparse,tkinter.tix"

[tool.pylint.'EXCEPTIONS']
overgeneral-exceptions= [
    "BaseException",
    "Exception"
]
